(Shading) Use Float32Arrays to store light data
Light data is passed to shaders as uniforms of known length, so there's
little need to store it in dynamically sized arrays on the CPU side. We
can use pre-allocated Float32Arrays of the same known length instead.

In each frame, these arrays are filled with zeros, and then position,
color and other data is written to them for each light present in the
scene.

This has the benefit of making it possible to use Light.Kind to break
out of the lighting loop in shaders, rather than pass the light_count
uniform explicitly. As soon as the light's kind is 0, we know we've
already processed all lights currently active in the scene.

(This is how lights are often handled in WebGL1 shaders, where it's
illegal to use uniform values in the loop's stop condition because the
spec allows specifications to statically unroll loops.)